Правки блока Модуляция
